"Refactor: 상점 운영시간 ZoneDateTime -> LocalTime -단순 시간정보만 표시 "
