+++ src/main/java/com/team5/catdogeats/users/service/impl/SellerInfoServiceImpl.java
@@ -4,8 +4,8 @@ import com.team5.catdogeats.users.domain.Users;
 import com.team5.catdogeats.users.domain.enums.DayOfWeek;
 import com.team5.catdogeats.users.domain.enums.Role;
 import com.team5.catdogeats.users.domain.mapping.Sellers;
-import com.team5.catdogeats.users.dto.SellerInfoRequest;
-import com.team5.catdogeats.users.dto.SellerInfoResponse;
+import com.team5.catdogeats.users.domain.dto.SellerInfoRequest;
+import com.team5.catdogeats.users.domain.dto.SellerInfoResponse;
 import com.team5.catdogeats.users.exception.BusinessNumberDuplicateException;
 import com.team5.catdogeats.users.exception.SellerAccessDeniedException;
 import com.team5.catdogeats.users.exception.UserNotFoundException;
